Adding the python frozen docs for the ColorSpacde class.
I was unable to build this locally so modified the CI workflow to upload the docs folder as an artifact in a separate branch. Even that was not readily consumable, I had to manually pick the differences and the remove the duplicate PyOpenColorIO in the entity names (i.e. PyOpenColorIO.PyOpenColorIO.ColorSpace)

Signed-off-by: cuneyt.ozdas <cuneyt.ozdas@autodesk.com>